+++ CHANGELOG.md
@@ -5,4 +5,4 @@ Make sure to clear this after each release
 Put changelog here:
 
 -----------------
-- Sound predicates now use suppliers instead of a predicate instance, fixing issues in multiplayer.
+- Fixed the Advancement API not properly adding requirements.